Revert "Convert Vhost Middleware Integration from JavaScript to TypeScript"
