+++ package.json
@@ -3,10 +3,9 @@
     "express": "^4.19.2",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1",
-    "vhost": "^3.0.2",
-    "vhost-ts": "^4.0.0"
+    "vhost": "^3.0.2"
   },
-  "type": "commonjs",
+  "type": "module",
   "name": "user",
   "version": "1.0.0",
   "main": "Index.js",
@@ -14,18 +13,9 @@
     "example": "example"
   },
   "scripts": {
-    "build": "tsc",
-    "start:user": "nodemon --exec ts-node user-subdomains.ts",
-    "start:virtual": "nodemon --exec ts-node virtual-domains.ts"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
-  "description": "",
-  "devDependencies": {
-    "@types/express": "^4.17.21",
-    "@types/vhost": "^3.0.9",
-    "nodemon": "^3.1.7",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.6.2"
-  }
+  "description": ""
 }
